requirements.txt
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
MSAL==1.13.0
azure-storage-blob==12.8.1
pyodbc==4.0.32
python-dotenv
